Extract error handling to its own file
